ai-python/src/celery_worker_hub/web_scraper/requirements.txt
beautifulsoup4==4.12.3
celery==5.4.0
langchain==0.3.6
langchain-community==0.3.5
langchain-openai==0.2.2
redis==5.0.3
flower==2.0.1
requests==2.32.3
fastapi==0.111.1
firebase-admin==6.5.0
pymongo==4.8.0
langchain-text-splitters==0.3.0
PyMuPDF==1.24.9
python-dotenv==1.0.1
pycryptodome==3.20.0
cryptography==42.0.5
jinja2==3.0.3
pydantic[email]>=2.0
boto3==1.34.60
botocore==1.34.60
Markdown==3.7
wdb==3.3.0
langchain-huggingface==0.1.0
langchain-core==0.3.19
huggingface-hub==0.26.2
langchain-anthropic==0.3.0